requirements.txt
mysql-connector>=2.2.9
slackclient==1.3.0
